データベースの追加・更新・取得ができるようになった

+++ main.py
@@ -9,9 +9,12 @@ __date__ = '2024/09/30 (Created: 2024/09/30)'
 FastAPIを用いたAPIの作成
 $ uvicorn main:app --reload
 でサーバーを起動
+
+APIのエンドポイント
+curl -X POST "http://127.0.0.1:8000/prompts/" -H  "accept: application/json" -H  "Content-Type: application/json" -d "{\"company_name\":\"asakura\",\"logo_image\":\"asakura.jpg\"}"
 """
 
-from fastapi import Depends, FastAPI
+from fastapi import Depends, FastAPI, HTTPException
 from sqlalchemy.orm import Session, sessionmaker
 from starlette.requests import Request
 from pydantic import BaseModel
@@ -22,12 +25,12 @@ SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 # Pydanticを用いたAPIに渡されるデータの定義 ValidationやDocumentationの機能が追加される
 class PromptIn(BaseModel):
-    title: str
-    done: bool
+    company_name: str
+    logo_image: str
 
 # 単一のPromptを取得するためのユーティリティ
 def get_prompt(db_session: Session, prompt_id: int):
-    return db_session.query(Prompt).filter(Prompt.id == prompt_id).first()
+    return db_session.query(Prompt).filter(Prompt.user_id == prompt_id).first()
 
 # DB接続のセッションを各エンドポイントの関数に渡す
 def get_db(request: Request):
@@ -39,45 +42,54 @@ app = FastAPI()
 # Promptの全取得
 @app.get("/prompts/")
 def read_prompts(db: Session = Depends(get_db)):
-    prompts = db.query(Prompt).all()
-    return prompts
+    return db.query(Prompt).all()
 
 # 単一のpromptを取得
 @app.get("/prompts/{prompt_id}")
 def read_prompt(prompt_id: int, db: Session = Depends(get_db)):
-    prompt = get_prompt(db, prompt_id)
-    return prompt
+    return get_prompt(db, prompt_id)
 
 # Promptを登録
-@app.post("/prompt/")
+@app.post("/prompts/")
 async def create_prompt(prompt_in: PromptIn,  db: Session = Depends(get_db)):
-    prompt = Prompt(title=prompt_in.title, done=False)
+    prompt = Prompt(**prompt_in.dict())
     db.add(prompt)
     db.commit()
-    prompt = get_prompt(db, prompt.id)
+    db.refresh(prompt)
     return prompt
 
 # promptを更新
 @app.put("/prompts/{prompt_id}")
 async def update_prompt(prompt_id: int, prompt_in: PromptIn, db: Session = Depends(get_db)):
+    # エラーハンドリング
+    if prompt is None:
+        raise HTTPException(status_code=404, detail="Prompt not found")
+
     prompt = get_prompt(db, prompt_id)
-    prompt.title = prompt_in.title
-    prompt.done = prompt_in.done
+    prompt.company_name = prompt_in.company_name
+    prompt.logo_image = prompt_in.logo_image
     db.commit()
-    prompt = get_prompt(db, prompt_id)
+    db.refresh(prompt)
     return prompt
 
 # promptを削除
 @app.delete("/prompts/{prompt_id}")
 async def delete_prompt(prompt_id: int, db: Session = Depends(get_db)):
     prompt = get_prompt(db, prompt_id)
+    # エラーハンドリング
+    if prompt is None:
+        raise HTTPException(status_code=404, detail="Prompt not found")
+    
     db.delete(prompt)
     db.commit()
 
 # リクエストの度に呼ばれるミドルウェア DB接続用のセッションインスタンスを作成
 @app.middleware("http")
 async def db_session_middleware(request: Request, call_next):
-    request.state.db = SessionLocal()
-    response = await call_next(request)
-    request.state.db.close()
+    response = None
+    try:
+        request.state.db = SessionLocal()
+        response = await call_next(request)
+    finally:
+        request.state.db.close()
     return response
